api/sms/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pika==1.2.0
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
twilio==7.14.1
urllib3==1.26.12
